osx: Fix a possible segfault in uv__io_poll
In our build infrastructure, I've seen a lot of segfaults recently that were all
only happening on OSX. Upon inspecting the coredumps, it appearded that all
segfaults happened at the same instruction, and upon translating the assembly
back to the source, I found that an array could be indexed with a -1 index
before the index was checked to be not -1.

I'm not 100% sure that this will fix our problems, but this check certainly
appears fishy to me.

XXX: this needs testing